Added firebase example and documentation
The main activity doesn't do anything at the moment, but it shows off how
to properly use firebase. Open up the firebase console and when you open up
the application you will see a new entry appear with the latest count. This basically
populates a new entry in the testArray on every load.