Wired listeners up to graphical components
Now our class has instance variables that are tied to actual interface
components that are in the layout. It also prevents the FAB from appearing
while the required attributes don't have any values. This makes it so that
only clean data can be added to firebase. Now we just have to actually use
firebase to push the data out.